Moving QueryAsync to Contrib project
